Connectors: move PlatformContext/Mcp.{ts,spec.ts} into Connectors/
Mechanical first step of the Connectors refactor (qhq-5d0.2 step 1):
the wire client moves out of the soon-to-be-deleted PlatformContext
folder into its new home. PlatformContext/index.ts re-points its imports
at ../Connectors/Mcp; tests still green. Subsequent steps build the
Connectors service around the wire client and delete PlatformContext
entirely.
