+++ catalog/app/components/Assistant/Model/Connectors/Mcp.spec.ts
@@ -32,7 +32,7 @@ const stubClient = (overrides: Partial<Mcp.McpClient> = {}): Mcp.McpClient => ({
   ...overrides,
 })
 
-describe('PlatformContext/Mcp', () => {
+describe('Connectors/Mcp', () => {
   describe('mapContent', () => {
     it('maps text blocks directly', () => {
       const out = Mcp.mapContent({ type: 'text', text: 'hello' })
